+++ schema/src/main/scala/net/reactivecore/cjs/restriction/NumberRestriction.scala
@@ -3,8 +3,8 @@ package net.reactivecore.cjs.restriction
 import io.circe.Codec
 import io.circe.generic.semiauto
 import net.reactivecore.cjs.util.Codecs
+import net.reactivecore.cjs.validator.ValidationProvider
 import net.reactivecore.cjs.validator.number._
-import net.reactivecore.cjs.validator.provider.{SequenceValidationProvider, ValidationProvider}
 
 case class NumberRestriction(
     minimum: OValidatingField[BigDecimal, MinimumValidator] = None,
@@ -18,5 +18,5 @@ object NumberRestriction {
   implicit lazy val codec: Codec.AsObject[NumberRestriction] = Codecs.withoutNulls(semiauto.deriveCodec)
 
   implicit lazy val validationProvider: ValidationProvider[NumberRestriction] =
-    SequenceValidationProvider[NumberRestriction]
+    ValidationProvider.visitingSequental[NumberRestriction]
 }
